+++ riskcontrol-spring-boot-starter/src/main/java/com/minlia/module/riskcontrol/endpoint/RiskIpListEndpoint.java
@@ -1,9 +1,16 @@
 package com.minlia.module.riskcontrol.endpoint;
 
+import com.alibaba.fastjson.JSONObject;
 import com.github.pagehelper.PageHelper;
 import com.github.pagehelper.PageInfo;
 import com.minlia.cloud.body.Response;
 import com.minlia.cloud.constant.ApiPrefix;
+import com.minlia.module.approved.bean.ro.ApprovedRO;
+import com.minlia.module.approved.constant.ApprovedSecurityConstant;
+import com.minlia.module.approved.entity.Approved;
+import com.minlia.module.approved.enumeration.ApprovedFunctionEnum;
+import com.minlia.module.approved.enumeration.ApprovedStatusEnum;
+import com.minlia.module.approved.service.ApprovedService;
 import com.minlia.module.audit.annotation.AuditLog;
 import com.minlia.module.audit.enumeration.OperationTypeEnum;
 import com.minlia.module.riskcontrol.bean.RiskIpListQRO;
@@ -16,6 +23,7 @@ import io.swagger.annotations.ApiOperation;
 import org.dozer.Mapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
@@ -34,12 +42,52 @@ public class RiskIpListEndpoint {
     @Autowired
     private RiskIpListMapper riskIpListMapper;
 
+    @Autowired
+    private ApprovedService approvedService;
+
+//    @AuditLog(value = "save fraud ip list", type = OperationTypeEnum.MODIFY)
+//    @PreAuthorize(value = "hasAnyAuthority('" + RiskSecurityConstants.IP_LIST_SAVE + "')")
+//    @ApiOperation(value = "保存")
+//    @PostMapping(value = "")
+//    public Response save(@Valid @RequestBody RiskIpList riskIpList) {
+//        riskIpListService.pub(riskIpList);
+//        return Response.success();
+//    }
     @AuditLog(value = "save fraud ip list", type = OperationTypeEnum.MODIFY)
     @PreAuthorize(value = "hasAnyAuthority('" + RiskSecurityConstants.IP_LIST_SAVE + "')")
     @ApiOperation(value = "保存")
     @PostMapping(value = "")
     public Response save(@Valid @RequestBody RiskIpList riskIpList) {
-        riskIpListService.pub(riskIpList);
+        Approved approved = Approved.builder()
+                .identifier(riskIpList.getCountry()+"&"+riskIpList.getStart())
+                .function(ApprovedFunctionEnum.RISK_MANAGE_IP_CONFIG_NEW_IP)
+                .afterData(JSONObject.toJSONString(riskIpList))
+                .build();
+        if(null != riskIpList.getId()){
+            RiskIpList oldRiskIpList = riskIpListService.queryById(riskIpList.getId());
+            if(null != oldRiskIpList){
+                approved.setIdentifier((oldRiskIpList.getCountry()+"&"+oldRiskIpList.getStart()));
+                approved.setBeforeData(JSONObject.toJSONString(oldRiskIpList));
+                approved.setFunction(ApprovedFunctionEnum.RISK_MANAGE_IP_CONFIG_EDIT);
+            }
+        }
+        approvedService.insert(approved);
+        return Response.success();
+    }
+
+    @AuditLog(value = "save fraud ip list", type = OperationTypeEnum.MODIFY)
+    @PreAuthorize(value = "hasAnyAuthority('" + ApprovedSecurityConstant.APPROVED_APPROVAL + "')")
+    @ApiOperation(value = "保存")
+    @PutMapping(value = "save/approval")
+    @Transactional
+    public Response save(@Valid @RequestBody ApprovedRO approvedRO) {
+        Approved approved = approvedService.approval(approvedRO);
+        if (ApprovedStatusEnum.APPROVED.equals(approvedRO.getStatus())) {
+            RiskIpList riskIpList = JSONObject.parseObject(approved.getAfterData(),RiskIpList.class);
+            riskIpListService.pub(riskIpList);
+            riskIpListService.updateCache();
+        }
+        approvedService.sendEmail(approved);
         return Response.success();
     }
 
@@ -52,12 +100,43 @@ public class RiskIpListEndpoint {
         return Response.success();
     }
 
+//    @AuditLog(value = "delete fraud ip list by id", type = OperationTypeEnum.DELETE)
+//    @PreAuthorize(value = "hasAnyAuthority('" + RiskSecurityConstants.IP_LIST_DELETE + "')")
+//    @ApiOperation(value = "ID删除")
+//    @DeleteMapping(value = "{id}")
+//    public Response delete(@PathVariable Long id) {
+//        riskIpListService.delete(id);
+//        return Response.success();
+//    }
     @AuditLog(value = "delete fraud ip list by id", type = OperationTypeEnum.DELETE)
     @PreAuthorize(value = "hasAnyAuthority('" + RiskSecurityConstants.IP_LIST_DELETE + "')")
     @ApiOperation(value = "ID删除")
     @DeleteMapping(value = "{id}")
     public Response delete(@PathVariable Long id) {
-        riskIpListService.delete(id);
+        RiskIpList riskIpList = riskIpListService.queryById(id);
+        Approved approved = Approved.builder()
+                .identifier(riskIpList.getCountry()+"&"+riskIpList.getStart())
+                .function(ApprovedFunctionEnum.RISK_MANAGE_IP_CONFIG_DELETE)
+                .beforeData(JSONObject.toJSONString(riskIpList))
+                .afterData(""+id)
+                .build();
+        approvedService.insert(approved);
+        return Response.success();
+    }
+
+    @AuditLog(value = "delete fraud ip list by id", type = OperationTypeEnum.DELETE)
+    @PreAuthorize(value = "hasAnyAuthority('" + ApprovedSecurityConstant.APPROVED_APPROVAL + "')")
+    @ApiOperation(value = "ID删除")
+    @PutMapping(value = "delete/approval")
+    @Transactional
+    public Response delete(@Valid @RequestBody ApprovedRO approvedRO) {
+        Approved approved = approvedService.approval(approvedRO);
+        if (ApprovedStatusEnum.APPROVED.equals(approvedRO.getStatus())) {
+            Long id = Long.valueOf(approved.getAfterData());
+            riskIpListService.delete(id);
+            riskIpListService.updateCache();
+        }
+        approvedService.sendEmail(approved);
         return Response.success();
     }
